pyproject.toml
[project]
name = "test-k"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "llama-index>=0.14.10",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "openai>=2.11.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.31.0",
    "python-decouple>=3.8",
    "mem0ai>=1.0.1",
    "rapidfuzz>=3.14.3",
    "twilio>=9.8.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"
